Code Cleanup
All session fixtures used by different tests are now in conftest.py
This is much cleaner and reduce code duplication
