+++ tests/fixtures/configWF2ui/core/roots.yml
@@ -41,7 +41,7 @@
 # core, this requirement has been lifted and the root can be named anything.
 # ------------------------------------------------------------------------------
 
-File Open UI Tests:
+WF2 UI Tests:
 
     description: A top-level root folder for production data. Project folders
         will be created beneath this location.
