+++ tests/fixtures/configWF2ui/core/schema/project.yml
@@ -12,4 +12,4 @@
 type: "project"
 
 # name of project root as defined in roots.yml
-root_name: "File Open UI Tests"
+root_name: "WF2 UI Tests"
